+++ src/main/java/com/xpanxion/Multiplication.java
@@ -9,14 +9,14 @@ public class Multiplication extends BinaryOperation {
     }
 
     @Override
-    protected int calculate(int multiplier, int multiplicand) {
+    protected long calculate(long multiplier, long multiplicand) {
 
         return multiplier * multiplicand;
 
     }
 
     @Override
-    protected boolean isValid(int multiplier, int multiplicand) {
+    protected boolean isValid(long multiplier, long multiplicand) {
 
         return true;
